bdgdiff: do not truncate region scores to integers
TwoConditionScores.mean_from_peakcontent declared the per-interval
score and its accumulator as `cython.long` after the pure-Python-mode
conversion (they were `float64_t` in MACS2 2.2.9.1 and MACS3 3.0.0), so
every log10 likelihood ratio was cut to its integer part before the
length-weighted mean was taken. With a cutoff below 1 every score in
the common file became 0 and cond1/cond2 scores could fall below the
cutoff (#715); with the default cutoff every score was biased low by
up to 1. Declare both as `cython.double` again. Restores the exact
scores in test/standard_results_bdgdiff.

Also fix the Test_TwoConditionScores fixture, which called the removed
bedGraphTrackI.safe_add_loc, and add a regression test.

Fixes #715

+++ MACS3/Signal/ScoreTrack.py
@@ -1935,8 +1935,8 @@ class TwoConditionScores:
         tmp_s: cython.int
         tmp_e: cython.int
         ln: cython.int
-        tmp_v: cython.long
-        sum_v: cython.long      # for better precision
+        tmp_v: cython.double
+        sum_v: cython.double    # for better precision
         r: cython.float
         i: cython.int
